+++ src/fides/api/api/v1/endpoints/privacy_request_endpoints.py
@@ -8,7 +8,6 @@ from datetime import datetime
 from typing import (
     Annotated,
     Any,
-    Callable,
     DefaultDict,
     Dict,
     List,
@@ -18,6 +17,16 @@ from typing import (
     Union,
 )
 
+from fides.api.api.deps import get_privacy_request_service
+from fides.services.messaging.messaging_service import (
+    check_and_dispatch_error_notifications,
+    send_verification_code_to_user,
+)
+from fides.services.privacy_request.privacy_request_service import (
+    PrivacyRequestService,
+    _trigger_pre_approval_webhooks,
+    queue_privacy_request,
+)
 import sqlalchemy
 from fastapi import Body, Depends, HTTPException, Security
 from fastapi.encoders import jsonable_encoder
@@ -49,12 +58,10 @@ from fides.api.api.v1.endpoints.manual_webhook_endpoints import (
 )
 from fides.api.common_exceptions import (
     FunctionalityNotConfigured,
-    IdentityNotFoundException,
     IdentityVerificationException,
     ManualWebhookFieldsUnset,
     MessageDispatchException,
     NoCachedManualWebhookEntry,
-    PolicyNotFoundException,
     TraversalError,
     ValidationError,
 )
@@ -75,7 +82,6 @@ from fides.api.models.privacy_preference import PrivacyPreferenceHistory
 from fides.api.models.privacy_request import (
     EXITED_EXECUTION_LOG_STATUSES,
     CheckpointActionRequired,
-    ConsentRequest,
     CustomPrivacyRequestField,
     ExecutionLog,
     ExecutionLogStatus,
@@ -84,7 +90,6 @@ from fides.api.models.privacy_request import (
     PrivacyRequestSource,
     PrivacyRequestStatus,
     ProvidedIdentity,
-    ProvidedIdentityType,
     RequestTask,
 )
 from fides.api.models.property import Property
@@ -97,10 +102,7 @@ from fides.api.oauth.utils import (
 from fides.api.schemas.dataset import CollectionAddressResponse, DryRunDatasetResponse
 from fides.api.schemas.external_https import PrivacyRequestResumeFormat
 from fides.api.schemas.messaging.messaging import (
-    FidesopsMessage,
     MessagingActionType,
-    RequestReceiptBodyParams,
-    RequestReviewDenyBodyParams,
 )
 from fides.api.schemas.policy import ActionType
 from fides.api.schemas.privacy_request import (
@@ -122,16 +124,9 @@ from fides.api.schemas.privacy_request import (
     ReviewPrivacyRequestIds,
     VerificationCode,
 )
-from fides.api.schemas.redis_cache import Identity
-from fides.api.service._verification import send_verification_code_to_user
 from fides.api.service.messaging.message_dispatch_service import (
     EMAIL_JOIN_STRING,
-    check_and_dispatch_error_notifications,
-    dispatch_message_task,
     message_send_enabled,
-)
-from fides.api.service.privacy_request.request_runner_service import (
-    queue_privacy_request,
 )
 from fides.api.service.privacy_request.request_service import (
     build_required_privacy_request_kwargs,
@@ -141,7 +136,6 @@ from fides.api.task.execute_request_tasks import log_task_queued, queue_request_
 from fides.api.task.filter_results import filter_data_categories
 from fides.api.task.graph_task import EMPTY_REQUEST, EMPTY_REQUEST_TASK, collect_queries
 from fides.api.task.task_resources import TaskResources
-from fides.api.tasks import MESSAGING_QUEUE_NAME
 from fides.api.util.api_router import APIRouter
 from fides.api.util.cache import FidesopsRedis
 from fides.api.util.collection_util import Row
@@ -190,6 +184,7 @@ from fides.common.api.v1.urn_registry import (
     REQUEST_TASK_CALLBACK,
     REQUEST_TASKS,
     V1_URL_PREFIX,
+    PRIVACY_REQUEST_RESUBMIT,
 )
 from fides.config import CONFIG
 from fides.config.config_proxy import ConfigProxy
@@ -229,8 +224,9 @@ def get_privacy_request_or_error(
 )
 def create_privacy_request(
     *,
-    db: Session = Depends(deps.get_db),
-    config_proxy: ConfigProxy = Depends(deps.get_config_proxy),
+    privacy_request_service: PrivacyRequestService = Depends(
+        deps.get_privacy_request_service
+    ),
     data: Annotated[List[PrivacyRequestCreate], Field(max_length=50)],  # type: ignore
 ) -> BulkPostPrivacyRequests:
     """
@@ -238,7 +234,9 @@ def create_privacy_request(
     or report failure and execute them within the Fidesops system.
     You cannot update privacy requests after they've been created.
     """
-    return create_privacy_request_func(db, config_proxy, data, authenticated=False)
+    return privacy_request_service.create_bulk_privacy_requests(
+        data, authenticated=False
+    )
 
 
 @router.post(
@@ -248,8 +246,9 @@ def create_privacy_request(
 )
 def create_privacy_request_authenticated(
     *,
-    db: Session = Depends(deps.get_db),
-    config_proxy: ConfigProxy = Depends(deps.get_config_proxy),
+    privacy_request_service: PrivacyRequestService = Depends(
+        get_privacy_request_service
+    ),
     client: ClientDetail = Security(
         verify_oauth_client,
         scopes=[PRIVACY_REQUEST_CREATE],
@@ -263,48 +262,8 @@ def create_privacy_request_authenticated(
     This route requires authentication instead of using verification codes.
     """
 
-    return create_privacy_request_func(
-        db, config_proxy, data, authenticated=True, user_id=client.user_id
-    )
-
-
-def _send_privacy_request_receipt_message_to_user(
-    policy: Optional[Policy],
-    to_identity: Optional[Identity],
-    service_type: Optional[str],
-    property_id: Optional[str],
-) -> None:
-    """Helper function to send request receipt message to the user"""
-    if not to_identity:
-        logger.error(
-            IdentityNotFoundException(
-                "Identity was not found, so request receipt message could not be sent."
-            )
-        )
-        return
-    if not policy:
-        logger.error(
-            PolicyNotFoundException(
-                "Policy was not found, so request receipt message could not be sent."
-            )
-        )
-        return
-    request_types: Set[str] = set()
-    for action_type in ActionType:
-        if policy.get_rules_for_action(action_type=ActionType(action_type)):
-            request_types.add(action_type)
-
-    dispatch_message_task.apply_async(
-        queue=MESSAGING_QUEUE_NAME,
-        kwargs={
-            "message_meta": FidesopsMessage(
-                action_type=MessagingActionType.PRIVACY_REQUEST_RECEIPT,
-                body_params=RequestReceiptBodyParams(request_types=request_types),
-            ).model_dump(mode="json"),
-            "service_type": service_type,
-            "to_identity": to_identity.labeled_dict(),
-            "property_id": property_id,
-        },
+    return privacy_request_service.create_bulk_privacy_requests(
+        data, authenticated=True, user_id=client.user_id
     )
 
 
@@ -1265,123 +1224,6 @@ def restart_privacy_request_from_failure(
     )
 
 
-def review_privacy_request(
-    db: Session,
-    request_ids: List[str],
-    process_request_function: Callable,
-    config_proxy: ConfigProxy,
-    webhook_id: Optional[str],
-    user_id: Optional[str],
-) -> BulkReviewResponse:
-    """Helper method shared between the approve and deny privacy request endpoints, and pre-approval webhook endpoints"""
-    succeeded: List[PrivacyRequest] = []
-    failed: List[Dict[str, Any]] = []
-
-    for request_id in request_ids:
-        privacy_request = PrivacyRequest.get(db, object_id=request_id)
-        if not privacy_request:
-            failed.append(
-                {
-                    "message": f"No privacy request found with id '{request_id}'",
-                    "data": {"privacy_request_id": request_id},
-                }
-            )
-            continue
-
-        if privacy_request.deleted_at is not None:
-            failed.append(
-                {
-                    "message": "Cannot transition status for a deleted request",
-                    "data": PrivacyRequestResponse.model_validate(
-                        privacy_request
-                    ).model_dump(mode="json"),
-                }
-            )
-            continue
-
-        if privacy_request.status != PrivacyRequestStatus.pending:
-            failed.append(
-                {
-                    "message": "Cannot transition status",
-                    "data": PrivacyRequestResponse.model_validate(
-                        privacy_request
-                    ).model_dump(mode="json"),
-                }
-            )
-            continue
-
-        try:
-            process_request_function(
-                db, config_proxy, privacy_request, webhook_id, user_id
-            )
-        except Exception:
-            failure = {
-                "message": "Privacy request could not be updated",
-                "data": PrivacyRequestResponse.model_validate(
-                    privacy_request
-                ).model_dump(mode="json"),
-            }
-            failed.append(failure)
-        else:
-            succeeded.append(privacy_request)
-
-    return BulkReviewResponse(
-        succeeded=succeeded,
-        failed=failed,
-    )
-
-
-def _send_privacy_request_review_message_to_user(
-    action_type: MessagingActionType,
-    identity_data: Dict[str, Any],
-    rejection_reason: Optional[str],
-    service_type: Optional[str],
-    property_id: Optional[str],
-) -> None:
-    """Helper method to send review notification message to user, shared between approve and deny"""
-    if not identity_data:
-        logger.error(
-            IdentityNotFoundException(
-                "Identity was not found, so request review message could not be sent."
-            )
-        )
-    to_identity: Identity = Identity(
-        email=identity_data.get(ProvidedIdentityType.email.value),
-        phone_number=identity_data.get(ProvidedIdentityType.phone_number.value),
-    )
-
-    dispatch_message_task.apply_async(
-        queue=MESSAGING_QUEUE_NAME,
-        kwargs={
-            "message_meta": FidesopsMessage(
-                action_type=action_type,
-                body_params=(
-                    RequestReviewDenyBodyParams(rejection_reason=rejection_reason)
-                    if action_type is MessagingActionType.PRIVACY_REQUEST_REVIEW_DENY
-                    else None
-                ),
-            ).model_dump(mode="json"),
-            "service_type": service_type,
-            "to_identity": to_identity.model_dump(mode="json"),
-            "property_id": property_id,
-        },
-    )
-
-
-def _trigger_pre_approval_webhooks(
-    db: Session, privacy_request: PrivacyRequest
-) -> None:
-    """
-    Shared method to trigger all configured pre-approval webhooks for a given privacy request.
-    """
-    pre_approval_webhooks = db.query(PreApprovalWebhook).all()
-    for webhook in pre_approval_webhooks:
-        privacy_request.trigger_pre_approval_webhook(
-            webhook=webhook,
-            policy_action=privacy_request.policy.get_action_type(),
-        )
-
-
 @router.post(
     PRIVACY_REQUEST_VERIFY_IDENTITY,
     status_code=HTTP_200_OK,
@@ -1448,58 +1290,6 @@ def verify_identification_code(
     return privacy_request  # type: ignore[return-value]
 
 
-def _approve_request(
-    db: Session,
-    config_proxy: ConfigProxy,
-    privacy_request: PrivacyRequest,
-    webhook_id: Optional[str],
-    user_id: Optional[str],
-) -> None:
-    """Method for how to process requests - approved"""
-    now = datetime.utcnow()
-    privacy_request.status = PrivacyRequestStatus.approved
-    privacy_request.reviewed_at = now
-    if user_id:
-        privacy_request.reviewed_by = user_id
-
-    if privacy_request.custom_fields:  # type: ignore[attr-defined]
-        privacy_request.custom_privacy_request_fields_approved_at = now
-        if user_id:
-            # for now, the reviewer will be marked as the approver of the custom privacy request fields
-            # this is to make it flexible in the future if we want to allow a different user to approve
-            privacy_request.custom_privacy_request_fields_approved_by = user_id
-    privacy_request.save(db=db)
-
-    auditlog_data = {
-        "privacy_request_id": privacy_request.id,
-        "action": AuditLogAction.approved,
-        "message": "",
-        "user_id": user_id if user_id else None,
-        "webhook_id": (
-            webhook_id if webhook_id else None
-        ),  # the last webhook reply received is what approves the entire request
-    }
-    AuditLog.create(
-        db=db,
-        data=auditlog_data,
-    )
-    if message_send_enabled(
-        db,
-        privacy_request.property_id,
-        MessagingActionType.PRIVACY_REQUEST_REVIEW_APPROVE,
-        config_proxy.notifications.send_request_review_notification,
-    ):
-        _send_privacy_request_review_message_to_user(
-            action_type=MessagingActionType.PRIVACY_REQUEST_REVIEW_APPROVE,
-            identity_data=privacy_request.get_cached_identity_data(),
-            rejection_reason=None,
-            service_type=config_proxy.notifications.notification_service_type,
-            property_id=privacy_request.property_id,
-        )
-
-    queue_privacy_request(privacy_request_id=privacy_request.id)
-
-
 @router.patch(
     PRIVACY_REQUEST_APPROVE,
     status_code=HTTP_200_OK,
@@ -1507,8 +1297,9 @@ def _approve_request(
 )
 def approve_privacy_request(
     *,
-    db: Session = Depends(deps.get_db),
-    config_proxy: ConfigProxy = Depends(deps.get_config_proxy),
+    privacy_request_service: PrivacyRequestService = Depends(
+        deps.get_privacy_request_service
+    ),
     client: ClientDetail = Security(
         verify_oauth_client,
         scopes=[PRIVACY_REQUEST_REVIEW],
@@ -1516,15 +1307,9 @@ def approve_privacy_request(
     privacy_requests: ReviewPrivacyRequestIds,
 ) -> BulkReviewResponse:
     """Approve and dispatch a list of privacy requests and/or report failure"""
-    user_id = client.user_id
 
-    return review_privacy_request(
-        db=db,
-        request_ids=privacy_requests.request_ids,
-        config_proxy=config_proxy,
-        process_request_function=_approve_request,
-        user_id=user_id,
-        webhook_id=None,
+    return privacy_request_service.approve_privacy_requests(
+        privacy_requests.request_ids, user_id=client.user_id
     )
 
 
@@ -1545,42 +1330,6 @@ def deny_privacy_request(
 ) -> BulkReviewResponse:
     """Deny a list of privacy requests and/or report failure"""
     user_id = client.user_id
-
-    def _deny_request(
-        db: Session,
-        config_proxy: ConfigProxy,
-        privacy_request: PrivacyRequest,
-        webhook_id: Optional[str],
-        user_id: Optional[str],
-    ) -> None:
-        """Method for how to process requests - denied"""
-        privacy_request.status = PrivacyRequestStatus.denied
-        privacy_request.reviewed_at = datetime.utcnow()
-        privacy_request.reviewed_by = user_id
-        privacy_request.save(db=db)
-        AuditLog.create(
-            db=db,
-            data={
-                "user_id": user_id,
-                "privacy_request_id": privacy_request.id,
-                "action": AuditLogAction.denied,
-                "message": privacy_requests.reason,
-            },
-        )
-
-        if message_send_enabled(
-            db,
-            privacy_request.property_id,
-            MessagingActionType.PRIVACY_REQUEST_REVIEW_DENY,
-            config_proxy.notifications.send_request_review_notification,
-        ):
-            _send_privacy_request_review_message_to_user(
-                action_type=MessagingActionType.PRIVACY_REQUEST_REVIEW_DENY,
-                identity_data=privacy_request.get_cached_identity_data(),
-                rejection_reason=privacy_requests.reason,
-                service_type=config_proxy.notifications.notification_service_type,
-                property_id=privacy_request.property_id,
-            )
 
     return review_privacy_request(
         db=db,
@@ -1617,7 +1366,9 @@ def mark_privacy_request_pre_approve_eligible(
     privacy_request_id: str,
     *,
     db: Session = Depends(deps.get_db),
-    config_proxy: ConfigProxy = Depends(deps.get_config_proxy),
+    privacy_request_service: PrivacyRequestService = Depends(
+        deps.get_privacy_request_service
+    ),
     webhook: PreApprovalWebhook = Security(
         verify_callback_oauth_pre_approval_webhook, scopes=[PRIVACY_REQUEST_REVIEW]
     ),
@@ -1694,13 +1445,8 @@ def mark_privacy_request_pre_approve_eligible(
         "All pre-approval webhooks have responded with eligible for privacy request '{}'. Proceeding to automatically approve request.",
         privacy_request_id,
     )
-    review_privacy_request(
-        db=db,
-        request_ids=[privacy_request_id],
-        config_proxy=config_proxy,
-        process_request_function=_approve_request,
-        webhook_id=webhook.id,
-        user_id=None,
+    privacy_request_service.approve_privacy_requests(
+        [privacy_request_id], webhook_id=webhook.id
     )
 
 
@@ -1848,7 +1594,7 @@ def privacy_request_data_transfer(
     db: Session = Depends(deps.get_db),
     cache: FidesopsRedis = Depends(deps.get_cache),
 ) -> Dict[str, Optional[List[Row]]]:
-    """Transfer access request iinformation to the parent server."""
+    """Transfer access request information to the parent server."""
     privacy_request = PrivacyRequest.get(db=db, object_id=privacy_request_id)
 
     if not privacy_request:
@@ -2090,241 +1836,6 @@ def resume_privacy_request_from_requires_input(
     )
 
     return privacy_request  # type: ignore[return-value]
-
-
-def create_privacy_request_func(
-    db: Session,
-    config_proxy: ConfigProxy,
-    data: Annotated[List[PrivacyRequestCreate], Field()],  # type: ignore
-    *,
-    authenticated: bool = False,
-    privacy_preferences: Optional[
-        List[PrivacyPreferenceHistory]
-    ] = None,  # For consent requests only
-    user_id: Optional[str] = None,
-) -> BulkPostPrivacyRequests:
-    """Creates privacy requests.
-
-    If authenticated is True the identity verification step is bypassed.
-    """
-    # TODO: (PROD-2142)- update privacy center to pass in property id where applicable
-    if not CONFIG.redis.enabled:
-        raise FunctionalityNotConfigured(
-            "Application redis cache required, but it is currently disabled! Please update your application configuration to enable integration with a redis cache."
-        )
-
-    privacy_preferences = privacy_preferences or []
-
-    created = []
-    failed = []
-    # Optional fields to validate here are those that are both nullable in the DB, and exist
-    # on the Pydantic schema
-
-    logger.info("Starting creation for {} privacy requests", len(data))
-
-    optional_fields = [
-        "external_id",
-        "started_processing_at",
-        "finished_processing_at",
-        "consent_preferences",
-        "property_id",
-        "source",
-    ]
-    for privacy_request_data in data:
-        if not any(privacy_request_data.identity.model_dump(mode="json").values()):
-            logger.warning(
-                "Create failed for privacy request with no identity provided"
-            )
-            failure = {
-                "message": "You must provide at least one identity to process",
-                "data": privacy_request_data.model_dump(mode="json"),
-            }
-            failed.append(failure)
-            continue
-
-        if privacy_request_data.property_id:
-            valid_property: Optional[Property] = Property.get_by(
-                db, field="id", value=privacy_request_data.property_id
-            )
-            if not valid_property:
-                logger.warning(
-                    "Create failed for privacy request with invalid property id"
-                )
-                failure = {
-                    "message": "Property id must be valid to process",
-                    "data": privacy_request_data.model_dump(mode="json"),
-                }
-                failed.append(failure)
-                continue
-
-        logger.info("Finding policy with key '{}'", privacy_request_data.policy_key)
-        policy: Optional[Policy] = Policy.get_by(
-            db=db,
-            field="key",
-            value=privacy_request_data.policy_key,
-        )
-        if policy is None:
-            logger.warning(
-                "Create failed for privacy request with invalid policy key {}'",
-                privacy_request_data.policy_key,
-            )
-
-            failure = {
-                "message": f"Policy with key {privacy_request_data.policy_key} does not exist",
-                "data": privacy_request_data.model_dump(mode="json"),
-            }
-            failed.append(failure)
-            continue
-
-        kwargs = build_required_privacy_request_kwargs(
-            privacy_request_data.requested_at,
-            policy.id,
-            config_proxy.execution.subject_identity_verification_required,
-            authenticated,
-        )
-        for field in optional_fields:
-            attr = getattr(privacy_request_data, field)
-            if attr is not None:
-                if field == "consent_preferences":
-                    attr = [consent.model_dump(mode="json") for consent in attr]
-
-                kwargs[field] = attr
-
-        # if the privacy request originated from the request manager (Admin UI)
-        # then add the user_id as the submitted_by user
-        if (
-            getattr(privacy_request_data, "source")
-            == PrivacyRequestSource.request_manager
-        ):
-            kwargs["submitted_by"] = user_id
-
-        try:
-            privacy_request: PrivacyRequest = PrivacyRequest.create(db=db, data=kwargs)
-            privacy_request.persist_identity(
-                db=db, identity=privacy_request_data.identity
-            )
-            _create_or_update_custom_fields(
-                db,
-                privacy_request,
-                privacy_request_data.consent_request_id,
-                privacy_request_data.custom_privacy_request_fields,
-            )
-            for privacy_preference in privacy_preferences:
-                privacy_preference.privacy_request_id = privacy_request.id
-                privacy_preference.save(db=db)
-
-            cache_data(
-                privacy_request,
-                policy,
-                privacy_request_data.identity,
-                privacy_request_data.encryption_key,
-                None,
-                privacy_request_data.custom_privacy_request_fields,
-            )
-
-            check_and_dispatch_error_notifications(db=db)
-
-            if not authenticated and message_send_enabled(
-                db,
-                privacy_request.property_id,
-                MessagingActionType.SUBJECT_IDENTITY_VERIFICATION,
-                config_proxy.execution.subject_identity_verification_required,
-            ):
-                send_verification_code_to_user(
-                    db,
-                    privacy_request,
-                    privacy_request_data.identity,
-                    privacy_request.property_id,
-                )
-                created.append(privacy_request)
-                continue  # Skip further processing for this privacy request
-
-            if not authenticated and message_send_enabled(
-                db,
-                privacy_request.property_id,
-                MessagingActionType.PRIVACY_REQUEST_RECEIPT,
-                config_proxy.notifications.send_request_receipt_notification,
-            ):
-                _send_privacy_request_receipt_message_to_user(
-                    policy,
-                    privacy_request_data.identity,
-                    config_proxy.notifications.notification_service_type,
-                    privacy_request.property_id,
-                )
-            if config_proxy.execution.require_manual_request_approval:
-                _trigger_pre_approval_webhooks(db, privacy_request)
-            else:
-                AuditLog.create(
-                    db=db,
-                    data={
-                        "user_id": "system",
-                        "privacy_request_id": privacy_request.id,
-                        "action": AuditLogAction.approved,
-                        "message": "",
-                    },
-                )
-                queue_privacy_request(privacy_request.id)
-
-        except MessageDispatchException as exc:
-            kwargs["privacy_request_id"] = privacy_request.id
-            logger.error("MessageDispatchException: {}", exc)
-            failure = {
-                "message": "Verification message could not be sent.",
-                "data": kwargs,
-            }
-            failed.append(failure)
-        except common_exceptions.RedisConnectionError as exc:
-            logger.error("RedisConnectionError: {}", Pii(str(exc)))
-            # Thrown when cache.ping() fails on cache connection retrieval
-            raise HTTPException(
-                status_code=HTTP_424_FAILED_DEPENDENCY,
-                detail=exc.args[0],
-            )
-        except Exception as exc:
-            as_string = Pii(str(exc))
-            error_cls = str(exc.__class__.__name__)
-            logger.error(f"Exception {error_cls}: {as_string}")
-            failure = {
-                "message": "This record could not be added",
-                "data": kwargs,
-            }
-            failed.append(failure)
-        else:
-            created.append(privacy_request)
-
-    # TODO: Don't return a 200 if there are failed requests, or at least not
-    # if there are zero successful ones
-    return BulkPostPrivacyRequests(
-        succeeded=created,
-        failed=failed,
-    )
-
-
-def _create_or_update_custom_fields(
-    db: Session,
-    privacy_request: PrivacyRequest,
-    consent_request_id: Optional[str],
-    custom_privacy_request_fields: Optional[Dict[str, Any]],
-) -> None:
-    """
-    Updates existing custom privacy request fields in the database with a privacy request ID.
-    Creates new custom privacy request fields if there aren't any available.
-
-    The presence or absence of custom fields is based on whether or not the creation of this
-    current privacy request was triggered by a consent request.
-    """
-    consent_request = ConsentRequest.get_by_key_or_id(
-        db=db, data={"id": consent_request_id}
-    )
-    if consent_request and consent_request.custom_fields:
-        for custom_field in consent_request.custom_fields:  # type: ignore[attr-defined]
-            custom_field.privacy_request_id = privacy_request.id
-            custom_field.save(db=db)
-    elif custom_privacy_request_fields:
-        privacy_request.persist_custom_privacy_request_fields(
-            db=db,
-            custom_privacy_request_fields=custom_privacy_request_fields,
-        )
 
 
 def _process_privacy_request_restart(
@@ -2673,3 +2184,27 @@ def get_test_privacy_request_results(
             else "DSR testing tools are not enabled, results will not be shown."
         ),
     }
+
+
+@router.post(
+    PRIVACY_REQUEST_RESUBMIT,
+    dependencies=[Security(verify_oauth_client, scopes=[PRIVACY_REQUEST_CREATE])],
+    response_model=PrivacyRequestResponse,
+)
+def resubmit_privacy_request(
+    privacy_request_id: str,
+    *,
+    privacy_request_service: PrivacyRequestService = Depends(
+        get_privacy_request_service
+    ),
+) -> PrivacyRequestResponse:
+    privacy_request = privacy_request_service.resubmit_privacy_request(
+        privacy_request_id
+    )
+
+    if not privacy_request:
+        raise HTTPException(
+            status_code=HTTP_404_NOT_FOUND, detail="Privacy request not found."
+        )
+
+    return privacy_request
