+++ src/fides/api/service/connectors/consent_email_connector.py
@@ -37,7 +37,6 @@ from fides.api.schemas.privacy_request import Consent
 from fides.api.schemas.redis_cache import Identity
 from fides.api.service.connectors.base_email_connector import (
     BaseEmailConnector,
-    get_email_messaging_config_service_type,
     get_org_name,
 )
 from fides.api.service.messaging.message_dispatch_service import dispatch_message
