+++ src/fides/api/util/cache.py
@@ -195,9 +195,7 @@ def get_masking_secret_cache_key(
 def get_all_cache_keys_for_privacy_request(privacy_request_id: str) -> List[Any]:
     """Returns all cache keys related to this privacy request's cached identities"""
     cache: FidesopsRedis = get_cache()
-    return cache.keys(f"{privacy_request_id}-*") + cache.keys(
-        f"id-{privacy_request_id}-*"
-    )
+    return cache.keys(f"*{privacy_request_id}*")
 
 
 def get_async_task_tracking_cache_key(privacy_request_id: str) -> str:
